🚚 move onboarding controller to module
